+++ apps/fluent-tester/src/theme/CustomThemes.ts
@@ -45,7 +45,7 @@ export class TesterThemeReference extends ThemeReference {
     super(
       baseTheme,
       (parent) => applyTheme(parent, this._themeName, this.options.appearance),
-      (parent) => applyBrand(parent, this._themeName, this._brand),
+      (parent) => applyBrand(parent, this._brand),
     );
     this.baseTheme = baseTheme;
     this.options = themeOptions;