+++ apps/fluent-tester/src/theme/ThemePickers.ios.tsx
@@ -1,4 +1,9 @@
 import { Button } from '@fluentui-react-native/experimental-button';
+import type { MenuAction } from '@react-native-menu/menu';
+import { MenuView } from '@react-native-menu/menu';
+import { testerTheme } from './CustomThemes';
+import type { ThemeNames } from './applyTheme';
+import { themeChoices } from './applyTheme';
 import type { Theme } from '@fluentui-react-native/framework';
 import { useTheme } from '@fluentui-react-native/framework';
 import { themedStyleSheet } from '@fluentui-react-native/themed-stylesheet';
@@ -39,10 +44,6 @@ export const themePickerStyles = themedStyleSheet((t: Theme) => {
 export const ThemePickers: React.FunctionComponent = () => {
   const themedStyles = themePickerStyles(useTheme());
 
-  const onBrandChange = React.useCallback((newBrand: string) => {
-    testerTheme.brand = newBrand as OfficeBrand;
-  }, []);
-
   const onThemeSelected = React.useCallback((newTheme: string) => {
     testerTheme.themeName = newTheme as ThemeNames;
   }, []);
@@ -51,12 +52,6 @@ export const ThemePickers: React.FunctionComponent = () => {
     id: themeChoice.value,
     title: themeChoice.label,
     state: testerTheme.themeName === themeChoice.value ? 'on' : 'off',
-  }));
-
-  const brandMenuOptions: MenuAction[] = brandOptions.map((brandOption) => ({
-    id: brandOption.value,
-    title: brandOption.label,
-    state: testerTheme.brand === brandOption.value ? 'on' : 'off',
   }));
 
   return (
@@ -69,15 +64,6 @@ export const ThemePickers: React.FunctionComponent = () => {
         actions={themeMenuOptions}
       >
         <Button appearance="subtle">Theme</Button>
-      </MenuView>
-      <MenuView
-        title="Brand"
-        onPressAction={({ nativeEvent }) => {
-          onBrandChange(nativeEvent.event);
-        }}
-        actions={brandMenuOptions}
-      >
-        <Button appearance="subtle">Brand</Button>
       </MenuView>
     </View>
   );