+++ packages/components/Button/src/ButtonColorTokens.android.ts
@@ -9,7 +9,7 @@ export const defaultButtonColorTokens: TokenSettings<ButtonTokens, Theme> = (t: 
      * 'primary', 'accent' and if no appearance is mentioned, picks this.
      */
     backgroundColor: t.colors.brandBackground,
-    rippleColor: 'transparent',
+    rippleColor: '#D4D4D4',
     color: t.colors.neutralForegroundOnColor,
     iconColor: t.colors.neutralForegroundOnColor,
     disabled: {
@@ -31,7 +31,7 @@ export const defaultButtonColorTokens: TokenSettings<ButtonTokens, Theme> = (t: 
     },
     subtle: {
       backgroundColor: 'transparent',
-      rippleColor: 'transparent',
+      rippleColor: '#D4D4D4',
       color: t.colors.brandForeground1,
       iconColor: t.colors.brandForeground1,
       disabled: {
@@ -54,7 +54,7 @@ export const defaultButtonColorTokens: TokenSettings<ButtonTokens, Theme> = (t: 
     },
     outline: {
       backgroundColor: 'transparent',
-      rippleColor: 'transparent',
+      rippleColor: '#D4D4D4',
       color: t.colors.brandForeground1,
       iconColor: t.colors.brandForeground1,
       borderColor: t.colors.brandStroke1,