+++ packages/components/Menu/src/MenuItem/MenuItemTokens.macos.ts
@@ -8,14 +8,14 @@ export const defaultMenuItemTokens: TokenSettings<MenuItemTokens, Theme> = (t: T
   backgroundColor: t.colors.transparentBackground,
   borderRadius: 5, // hardcoded for now to match NSMenu
   checkmarkSize: 16,
-  color: t.colors.menuItemText, // hardcoded for now to match NSMenu
+  color: t.colors.neutralForeground1,
   fontFamily: t.typography.families.primary,
   fontSize: 13, // aligning with NSMenu font size
   fontWeight: globalTokens.font.weight.regular as FontWeightValue,
   gap: globalTokens.size40,
   paddingHorizontal: 5, // hardcoded for now to match NSMenu
   paddingVertical: 3, // hardcoded for now to match NSMenu
-  submenuIndicatorColor: t.colors.neutralForeground2,
+  submenuIndicatorColor: t.colors.neutralForeground1,
   submenuIndicatorPadding: globalTokens.sizeNone,
   submenuIndicatorSize: 16,
   focused: {