+++ packages/utils/tokens/src/text-tokens.ts
@@ -15,16 +15,32 @@ export interface FontStyleTokens {
   fontWeight?: keyof Typography['weights'] | TextStyle['fontWeight'];
   fontLineHeight?: TextStyle['lineHeight'];
   fontLetterSpacing?: TextStyle['letterSpacing'];
+  fontStyle?: TextStyle['fontStyle'];
   // Props below are used on iOS only. GH #2268: Import these from RN directly
   fontDynamicTypeRamp?: string;
   fontMaximumSize?: number;
 }
 
-export type FontTokens = FontStyleTokens & FontVariantTokens;
+export interface FontDecorationTokens {
+  textDecorationLine?: TextStyle['textDecorationLine'];
+}
+
+export type FontTokens = FontStyleTokens & FontVariantTokens & FontDecorationTokens;
 
 export const fontStyles: TokenBuilder<FontTokens> = {
   from: (
-    { fontDynamicTypeRamp, fontFamily, fontLetterSpacing, fontLineHeight, fontMaximumSize, fontSize, fontWeight, variant }: FontTokens,
+    {
+      fontDynamicTypeRamp,
+      fontFamily,
+      fontLetterSpacing,
+      fontLineHeight,
+      fontMaximumSize,
+      fontSize,
+      fontStyle,
+      fontWeight,
+      textDecorationLine,
+      variant,
+    }: FontTokens,
     { typography }: Theme,
   ) => {
     const { families, sizes, weights, variants } = typography;
@@ -41,11 +57,13 @@ export const fontStyles: TokenBuilder<FontTokens> = {
       return {
         fontFamily: families[fontFamily] ?? fontFamily ?? families[variants[variant]?.face] ?? variants[variant]?.face,
         fontSize: sizes[fontSize] ?? fontSize ?? sizes[variants[variant]?.size] ?? variants[variant]?.size,
+        fontStyle: fontStyle,
         fontWeight: weights[fontWeight] ?? fontWeight ?? weights[variants[variant]?.weight] ?? variants[variant]?.weight,
         lineHeight: fontLineHeight ?? variants[variant]?.lineHeight,
         letterSpacing: fontLetterSpacing ?? variants[variant]?.letterSpacing,
         dynamicTypeRamp: fontDynamicTypeRamp ?? variants[variant]?.dynamicTypeRamp,
         maximumFontSize: fontMaximumSize,
+        textDecorationLine,
       };
     }
 
@@ -58,8 +76,10 @@ export const fontStyles: TokenBuilder<FontTokens> = {
     'fontLetterSpacing',
     'fontMaximumSize',
     'fontSize',
+    'fontStyle',
     'fontWeight',
     'variant',
+    'textDecorationLine',
   ],
 };