+++ packages/components/Link/CHANGELOG.md
@@ -1,8 +1,21 @@
 # Change Log - @fluentui-react-native/link
 
-This log was last generated on Sat, 11 Feb 2023 01:32:30 GMT and should not be manually modified.
+This log was last generated on Fri, 24 Feb 2023 02:22:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.18.16
+
+Fri, 24 Feb 2023 02:22:40 GMT
+
+### Patches
+
+- Fix underline styling not applying as a result of FontToken changes (winlarry@microsoft.com)
+- Bump @uifabricshared/foundation-compose to v1.13.2
+- Bump @fluentui-react-native/framework to v0.9.2
+- Bump @fluentui-react-native/interactive-hooks to v0.22.19
+- Bump @fluentui-react-native/text to v0.19.21
+- Bump @fluentui-react-native/tokens to v0.20.9
 
 ## 0.18.15