+++ packages/experimental/NativeDatePicker/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @fluentui-react-native/experimental-native-date-picker
 
-This log was last generated on Thu, 02 Feb 2023 01:29:38 GMT and should not be manually modified.
+This log was last generated on Fri, 17 Feb 2023 06:30:14 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.6
+
+Fri, 17 Feb 2023 06:30:14 GMT
+
+### Patches
+
+- Update native modules to FUA 0.13 (mischreiber@microsoft.com)
 
 ## 0.7.5