+++ packages/experimental/Spinner/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/spinner
 
-This log was last generated on Sat, 11 Feb 2023 01:32:30 GMT and should not be manually modified.
+This log was last generated on Fri, 24 Feb 2023 02:22:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.2
+
+Fri, 24 Feb 2023 02:22:40 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.9.2
+- Bump @fluentui-react-native/text to v0.19.21
 
 ## 0.2.1