+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Sat, 11 Feb 2023 01:32:30 GMT and should not be manually modified.
+This log was last generated on Fri, 24 Feb 2023 02:22:40 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.25.2
+
+Fri, 24 Feb 2023 02:22:40 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.17.2
+- Bump @fluentui-react-native/theming-utils to v0.23.4
 
 ## 0.25.1